requirements.txt
scikit-learn
pandas
nltk
openai
python-dotenv
streamlit[openai]
gdown